% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addData.R
\name{addData}
\alias{addData}
\title{private: Insert data into the EMEP database (please use addFile)}
\format{R Code}
\usage{
addData(emepdata, file_id_IN, sgl_id_IN)
}
\arguments{
\item{emepdata}{list: EMEP data to write into the database}

\item{file_id_IN}{integer: id of the entry in the files table corresponding to the currently being imported data set}

\item{sgl_id_IN}{integer: id of the entry in the sgl_files table corresponding to the currently being imported data set}
}
\value{
list(): information whether the process was successful or not
}
\description{
Please use addFile() to add data to the database if you 
are not really sure what you are doing. Adds data (emepdata) 
into the database defined by the DBaccess variables. A 
RODBC connection (channel) is opend for this purpose.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

