% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addData.RODBC.R
\name{addData.RODBC}
\alias{addData.RODBC}
\title{private: Insert data into the EMEP database (please use addFile)}
\format{R Code
We have three possible situations:
 (a) Entry in f_substance exists and is complete (sf_remove = FALSE, sf_insert = FALSE)
       - do nothing
 (b) Entry in f_substance exists and is not complete (sf_remove = TRUE, sf_insert = TRUE)
       - remove rows in data table which have the corresponding sf_id
       - insert the new data (if no error arose during removal)
 (c) Entry in f_substance exists not (sf_remove = FALSE, sf_insert = TRUE)
       - insert the data}
\usage{
addData.RODBC(channel, emepdata, file_id_IN, sgl_id_IN)
}
\arguments{
\item{channel}{RODBC: open ODBC data basebase connection}

\item{emepdata}{list: EMEP data to write into the database}

\item{file_id_IN}{integer: id of the entry in the files table corresponding to the currently being imported data set}

\item{sgl_id_IN}{integer: id of the entry in the sgl_files table corresponding to the currently being imported data set}
}
\value{
list(): information whether the process was successful or not
}
\description{
Please use addFile() to add data to the database if you 
are not really sure what you are doing. Adds data (emepdata) 
into the database given by a open RODBC connection 
(channel).
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

