% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFile.RODBC.R
\name{addFile.RODBC}
\alias{addFile.RODBC}
\title{Insert data into the EMEP database}
\format{R Code
There are three different cases:
 (a) file was never imported before
 (b) file has benn imported but not completely
 (c) file has been completely imported 
In the case (c) we have to do nothing (noInsert = TRUE).
In the cases (a) and (b) we have to proceed. But in case (a) we have to 
create an database entry for the file, first.}
\usage{
addFile.RODBC(channel, file, removeDuplicates = TRUE,
  removeStatistics = TRUE, forceCleanAndInsertProblematicFiles = FALSE)
}
\arguments{
\item{channel}{RODBC: open ODBC data basebase connection}

\item{file}{character: path of the EMEP text file to import (path = dir/file)}

\item{removeDuplicates}{logical [TRUE]: the function removeDuplicates.emepdata() is applied on the emepdata before insertion}

\item{removeStatistics}{logical [TRUE]: the function removeStatistics.emepdata() is applied on the emepdata before insertion (variables with 'Statistics' attributes != 'arithmetic mean' will be removed)}

\item{forceCleanAndInsertProblematicFiles}{logical [FALSE]: If TRUE, problematic variables in files are removed and than the variables are inserted into the database.}
}
\value{
list(): information whether the process was successful or not
}
\description{
This function reads in a EMEP text file 'file', which 
needs to be in the EBAS v1.1 format. The data will be written into the 
database to which an open RODBC connection exists given by 'channel'.
}
\details{
This function reads in a EMEP text file 'file', which 
needs to be in the EBAS v1.1 format. The data will be written into the 
database defined by the DBaccess variable. A RODBC connection ('channel')
is opened and addFile.RODBC is called.

If forceCleanAndInsertProblematicFiles is FALSE, some files will not be inserted (when there 
is at least one problematic variable present). This is tested by 
'findProblematicAttribute.emepdata()'. See the documentation of that function
to see, which variables with which attributes might be problematic. If 
forceCleanAndInsertProblematicFiles is set to TRUE, the problematic variables will be removed
and the remaining variables will be inserted. The file will be marked as inserted incomplete.

Note:
If the file is not in version 1.1 of the EBAS format but in another version 
please test the read.emepFile(...) individually on that file.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

