% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFlag.R
\name{addFlag}
\alias{addFlag}
\title{Adds a new numflag to the EMEP database}
\format{R Code}
\usage{
addFlag(flag, create_group = FALSE, no_insert_on_warning = TRUE)
}
\arguments{
\item{flag}{list(): list with information on the flag to insert; flaggroup is optional if floor(numflag/100)==flaggroup; group_description is optional if no new flaggroup should be created}

\item{create_group}{logical: create a new flaggroup if necessary (default = FALSE)}

\item{no_insert_on_warning}{logical: TRUE = do nothing if a warning was thrown (default = TRUE)}
}
\value{
list(err_code, err_string) list(integer, character): err_code (0 if successful, 1 if not successfull)
}
\description{
Adds a new numflag to the EMEP database
}
\details{
Structure of flag: list(numflag, status, description, flaggroup, group_description)
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

