% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFSubstance.R
\name{addFSubstance}
\alias{addFSubstance}
\title{private: adds and new row into the f_substance table}
\format{R Code}
\usage{
addFSubstance(sgl_id, substance_short, matrix, method_ref)
}
\arguments{
\item{sgl_id}{integer: sgl_id pointing to a row in the sgl_files tables}

\item{substance_short}{character: short substance name}

\item{matrix}{character: matrix of the measurement (air, aerosol, precip)}

\item{method_ref}{character: method_ref of the measurement}
}
\value{
list(err, exists, complete, sf_id) list(integer, logical, logical, integer): exists or not
}
\description{
Inserts a new line into the f_substances table. For this purpose and a
connection to the database defined by the variable DBaccess is established.
The function addFSubstance.RODBC does the actual work.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

