% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFSubstance.RODBC.R
\name{addFSubstance.RODBC}
\alias{addFSubstance.RODBC}
\title{private: adds and new row into the f_substance table}
\format{R Code}
\usage{
addFSubstance.RODBC(channel, sgl_id, substance_short, matrix, method_ref)
}
\arguments{
\item{channel}{RODBC: open RODBC database connection}

\item{sgl_id}{integer: sgl_id pointing to a row in the sgl_files tables}

\item{substance_short}{character: short substance name}

\item{matrix}{character: matrix of the measurement (air, aerosol, precip)}

\item{method_ref}{character: method_ref of the measurement}
}
\value{
list(err, exists, complete, sf_id) list(integer, logical, logical, integer): exists or not
}
\description{
Inserts a new line into the f_substances table of the database,
to which an open connection via 'channel' exists.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

