% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStation.RODBC.R
\name{addStation.RODBC}
\alias{addStation.RODBC}
\title{Adds a new station to the EMEP database}
\format{R Code}
\usage{
addStation.RODBC(channel, station)
}
\arguments{
\item{channel}{RODBC: open RODBC database connection}

\item{station}{list(station_id, station_name, country, longitude, latitude, altitude): list with information on the station to insert; station_id needs to be a 7-character string}
}
\value{
list(err_code, err_string) list(integer, character): err_code (0 if successful, 1 if not successfull)
}
\description{
Adds a new station to the EMEP database
}
\details{
The structure of 'station' should be: list(station_id, station_name, country, longitude, latitude, altitude)
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

