% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeFSubstance.R
\name{completeFSubstance}
\alias{completeFSubstance}
\title{private: set a row in the f_substance table as 'complete'}
\format{R Code}
\usage{
completeFSubstance(sf_id)
}
\arguments{
\item{sf_id}{integer: id of an entry in the f_substances table}
}
\value{
list(err, exists, complete, sf_id): list containing one integer
         - indicating if no error (==0) or an error (!=0) occured -, three 
         booleans - indicating if the endtry already exists and if it is
         complete - and the sf_id which was parsed at the beginning.
}
\description{
Sets the 'complete' boolean of one entry in table 'f_substances'
to TRUE. Returns whether it was successful or not. Does not test
whether an entry exists, in advance. A connection to 
the database defined by the variable DBaccess is established.
The function existsFSubstance.RODBC does the actual work.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

