% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsFSubstance.R
\name{existsFSubstance}
\alias{existsFSubstance}
\title{private: checks whether a certain variable already exists in the f_substance table}
\format{R Code}
\usage{
existsFSubstance(sgl_id, substance_short, matrix, method_ref)
}
\arguments{
\item{sgl_id}{integer: sgl_id pointing to a row in the sgl_files tables}

\item{substance_short}{character: short substance name}

\item{matrix}{character: matrix of the measurement (air, aerosol, precip)}

\item{method_ref}{character: method_ref of the measurement}
}
\value{
list(err, exists, complete, sf_id) list(integer, logical, logical, integer): exists or not
}
\description{
Tests whether a substance measured in a given matrix and method_ref 
(pointing to sgl_id) does alreay exist in the database. A connection to 
the database defined by the variable DBaccess is established.
The function existsFSubstance.RODBC does the actual work.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

