% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{removeData}
\alias{removeData}
\title{private: remove rows from data table}
\format{R Code}
\usage{
removeData(sf_id)
}
\arguments{
\item{sf_id}{integer: id of an entry in the f_substances table}
}
\value{
list(err_code, n_rows) list(integer, integer): !=0 if an error occured; number of removed rows
}
\description{
Removes all rows from the data table which have a given sf_id. Prior to 
sending the DELETE query, the presence of data row with this sf_id is
tested.
}
\author{
Daniel Neumann, daniel.neumann@___.de
}

