% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.RODBC.R
\name{removeData.RODBC}
\alias{removeData.RODBC}
\title{private: remove rows from data table}
\format{R Code}
\usage{
removeData.RODBC(channel, sf_id)
}
\arguments{
\item{channel}{RODBC: open ODBC data basebase connection}

\item{sf_id}{integer: id of an entry in the f_substances table}
}
\value{
list(err_code, n_rows) list(integer, integer): !=0 if an error occured; number of removed rows
}
\description{
Removes all rows from the data table which have a given sf_id. Prior to 
sending the DELETE query, the presence of data row with this sf_id is
tested.
}
\author{
Daniel Neumann, daniel.neumann@___.de
}

