% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeStation.RODBC.R
\name{removeStation.RODBC}
\alias{removeStation.RODBC}
\title{Removes a station from the EMEP database}
\format{R Code}
\usage{
removeStation.RODBC(channel, station_id = "", station_name = "")
}
\arguments{
\item{channel}{RODBC: open RODBC database connection}

\item{station_id}{character: station id of the station to remove; is unambiguous; is optional if station_name is set;}

\item{station_name}{character (optional): name of the station(s) to remove; might be ambiguous; obligatory if no station_id is set;}
}
\value{
list(err_code, err_string) list(integer, character): err_code (0 if successful, 1 if not successfull)
}
\description{
Removes a station from the EMEP database
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

